pyproject.toml
[project]
name = "agentic-chat"
version = "0.1.0"
description = "Agentic terminal and web chat app with OpenRouter + tool calling."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "exa-py>=2.10.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.55.0",
    "ujson>=5.10.0",
]

[project.scripts]
agentic-chat = "agentic_chat.app.entrypoint:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.7",
]
